+++ src/main/java/io/github/sakurawald/module/mixin/chat/stripe/ServerPlayNetworkHandlerMixin.java
@@ -2,6 +2,7 @@ package io.github.sakurawald.module.mixin.chat.stripe;
 
 import io.github.sakurawald.core.auxiliary.LogUtil;
 import io.github.sakurawald.core.service.style_striper.StyleStriper;
+import io.github.sakurawald.module.initializer.chat.style.ChatStyleInitializer;
 import net.minecraft.network.packet.c2s.play.ChatMessageC2SPacket;
 import net.minecraft.server.network.ServerPlayNetworkHandler;
 import net.minecraft.server.network.ServerPlayerEntity;
@@ -11,7 +12,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.ModifyVariable;
 
 // the pipeline of onChatMessage() should be: chat.stripe -> chat.mention -> chat.rewrite
-@Mixin(value = ServerPlayNetworkHandler.class, priority = 1000 + 500)
+@Mixin(value = ServerPlayNetworkHandler.class, priority = 1000 + 750)
 public abstract class ServerPlayNetworkHandlerMixin {
 
     @Shadow
@@ -20,7 +21,7 @@ public abstract class ServerPlayNetworkHandlerMixin {
     @ModifyVariable(method = "onChatMessage", at = @At(value = "HEAD"), argsOnly = true)
     public ChatMessageC2SPacket modifyChatMessageSentByPlayers(ChatMessageC2SPacket original) {
         String oldChatMessage = original.chatMessage();
-        String newChatMessage = StyleStriper.stripe(player, StyleStriper.STYLE_TYPE_CHAT, oldChatMessage);
+        String newChatMessage = StyleStriper.stripe(player, ChatStyleInitializer.STYLE_TYPE_CHAT, oldChatMessage);
         LogUtil.debug("stripe chat message: old = {}, new = {}", oldChatMessage, newChatMessage);
 
         return new ChatMessageC2SPacket(newChatMessage, original.timestamp(), original.comp_947(), original.comp_948(), original.acknowledgment());