+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 > The version number of fuji follows `semver` now: https://semver.org/
 
-- (back module) feature: persist the `saved-location` for `/back` command.
+- (chat.style module) fix: when `chat.stripe` module is disabled, the `/chat style set <style>` command should not stripe the style tags.
+- (chat.replace module) fix: the sender and receiver is wrong, it should parse the message as sender, not the receiver.